pyproject.toml
[tool.poetry]
name = "airflow-app"
version = "0.1.0"
description = "Airflow 2.10.3 application"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
apache-airflow = "2.10.3"
apache-airflow-providers-postgres = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
